requirements.txt
fastapi
uvicorn[standard]
openai
pdfplumber
python-multipart
firebase-admin
aiosmtplib
python-dotenv
reportlab
